M00/EX03
Added a 2nd part that introduces the `__main__` concept. It should make the correction easier without requiring an errata.

Removed a sneaky extra-space in the example

Replaced python with python3: less confusing with default envs.

The docstring is now mandatory. It is no longer required to handle multiple variable in the function (*args is in M2)

The function must print an error when a non-string is given.